[Misc] Code changes
Created a new REST resource to update the supporters or opponents of a given Idea.
Refactored the IdeasManager code to reduce code duplication.
Separated the REST model from the domain Model. Created a Utility class to convert between the two.
Added a new public method to the IdeaManager for the removal of votes.