Investigate the need of programming rights #46 (#47)
* Investigate the need of programming rights #46

* Code changes

Migrated the code from IdeasVoteService to Java. Modified the JavaScript extension to make the call to the correct URL. Modified all the files that required the JS code.

* Added unit tests for DefaultIdeasManager

* [Misc] Code improvements

* [Misc] Code improvements

* [Misc] Code changes

* [Misc] Code changes

Created a new REST resource to update the supporters or opponents of a given Idea.
Refactored the IdeasManager code to reduce code duplication.
Separated the REST model from the domain Model. Created a Utility class to convert between the two.
Added a new public method to the IdeaManager for the removal of votes.

* [Misc] Code changes

* Deleted IdeasVoteResource but inregrated its functions in IdeasResource
* Use query parameter to determine if the vote is for or against
* Removed parameter from removeVote functions

* [Misc] Code improvements

* Changed the page comment on save to be more accurate
* Changed the Idea model slightly
* Removed the public acces modifier of the DefaultIdeasManagerTest class
* Fixed IdeasResource test
* When voting, used one request to the server rather than two

* [Misc] Upgrade Licensor dependency to 1.22.1

* [Misc] Code Changes

* Escaped translated strings in JS